+++ strutil/fuzz.go
@@ -1,0 +1,16 @@
+//go:build gofuzz
+// +build gofuzz
+
+package strutil
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+//                                                                                    //
+//                         Copyright (c) 2024 ESSENTIAL KAOS                          //
+//      Apache License, Version 2.0 <https://www.apache.org/licenses/LICENSE-2.0>     //
+//                                                                                    //
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+func Fuzz(data []byte) int {
+	Fields(string(data))
+	return 0
+}