+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * `[req]` Make `Limiter` public
 * `[log]` Added `WithFullCallerPath` option to enable the output of the full caller path
 * `[strutil]` Added support of escaped strings to `Fields`
+* `[strutil]` Added fuzz tests for `Fields` method
 
 ### [13.11.0](https://kaos.sh/ek/13.11.0)